Use consolidated omeda+identityx setup
