+++ packages/shared/package.json
@@ -21,10 +21,10 @@
     "@parameter1/base-cms-marko-web-gtm": "^2.48.1",
     "@parameter1/base-cms-marko-web-html-sitemap": "^2.46.1",
     "@parameter1/base-cms-marko-web-icons": "^2.46.3",
-    "@parameter1/base-cms-marko-web-identity-x": "^2.52.0",
+    "@parameter1/base-cms-marko-web-identity-x": "^2.56.0",
     "@parameter1/base-cms-marko-web-native-x": "^2.48.1",
-    "@parameter1/base-cms-marko-web-omeda": "^2.52.0",
-    "@parameter1/base-cms-marko-web-omeda-identity-x": "^2.52.0",
+    "@parameter1/base-cms-marko-web-omeda": "^2.54.0",
+    "@parameter1/base-cms-marko-web-omeda-identity-x": "^2.56.0",
     "@parameter1/base-cms-marko-web-p1-events": "^2.52.0",
     "@parameter1/base-cms-marko-web-photoswipe": "^2.45.0",
     "@parameter1/base-cms-marko-web-reveal-ad": "^2.45.0",
