+++ packages/shared/templates/user/index.js
@@ -1,0 +1,13 @@
+const authenticate = require('./authenticate');
+const login = require('./login');
+const logout = require('./logout');
+const register = require('./register');
+const profile = require('./profile');
+
+module.exports = {
+  authenticate,
+  login,
+  logout,
+  register,
+  profile,
+};
